refactoring the code more
